Moved path function to utilities
